Rewritten affecting things by wind spell.

+++ src/engine_camera.c
@@ -64,8 +64,8 @@ MapCoordDelta get_3d_box_distance(const struct Coord3d *pos1, const struct Coord
 MapCoordDelta get_2d_box_distance(const struct Coord3d *pos1, const struct Coord3d *pos2)
 {
   long dist_x,dist_y;
-  dist_y = abs(pos1->y.val - (long)pos2->y.val);
-  dist_x = abs(pos1->x.val - (long)pos2->x.val);
+  dist_y = abs((long)pos1->y.val - (long)pos2->y.val);
+  dist_x = abs((long)pos1->x.val - (long)pos2->x.val);
   if (dist_y <= dist_x)
     return dist_x;
   return dist_y;
